+++ src/fast5_rekindler/cleanup.py
@@ -1,13 +1,15 @@
-import h5py
-from h5py import Group
+from typing import Any
 
-def delete_configuration_groups(group: Group) -> None:
+import h5py  # type: ignore
+
+
+def delete_configuration_groups(group: Any) -> None:
     """
     Recursively delete groups named 'Configuration'.
-    
+
     Params:
         group (h5py.Group): A group object.
-        
+
     Returns:
         None
     """
@@ -17,23 +19,24 @@ def delete_configuration_groups(group: Group) -> None:
         item = group[key]
         if isinstance(item, h5py.Group):
             # If the item is a group and its name is 'Configuration', delete it
-            if key == 'Configuration':
+            if key == "Configuration":
                 del group[key]
             else:
                 # Otherwise, recursively check its subgroups
                 delete_configuration_groups(item)
+
 
 def cleanup_fast5(fast5_filepath: str) -> None:
     """
     Delete configuration groups from a FAST5 file.
     These groups are added automatically by ont-fast5-api
     but are not needed in rekindled FAST5 files.
-    
+
     Params:
         fast5_filepath (str): Path to a FAST5 file.
-        
+
     Returns:
         None
     """
-    with h5py.File(fast5_filepath, 'r+') as file:
+    with h5py.File(fast5_filepath, "r+") as file:
         delete_configuration_groups(file)