+++ CHANGELOG.md
@@ -4,5 +4,5 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-- First functional release of code
-- Collates information in POD5 and BAM file into legacy basecalled FAST5 files
+### Added
+- First fully functional release of code which can collate info in POD5 and BAM file into legacy basecalled FAST5 files